+++ src/index.ts
@@ -46,6 +46,12 @@ export {
 } from './lib/provable-types/merkle-list.js';
 
 export * as Mina from './lib/mina.js';
+export {
+  type Transaction,
+  type PendingTransaction,
+  type IncludedTransaction,
+  type RejectedTransaction,
+} from './lib/mina/transaction.js';
 export type { DeployArgs } from './lib/zkapp.js';
 export {
   SmartContract,
